дз 5го раздела готова к проверке
